+++ src/main/java/de/uros/citlab/module/baseline2polygon/Baseline2PolygonParser.java
@@ -207,7 +207,7 @@ public class Baseline2PolygonParser implements IBaseline2Polygon {
 
     @Override
     public String getProvider() {
-        return MetadataUtil.getProvider("Tobias Gruening", "tobias.gruening@uni-rostock.de");
+        return MetadataUtil.getProvider("Tobias Gruening", "tobias.gruening@planet-ai.de");
     }
 
 }