+++ src/main/java/de/uros/citlab/module/feat/FeatHannFreq.java
@@ -92,7 +92,7 @@ public class FeatHannFreq implements IFeatureGeneratorStreamable {
 
     @Override
     public String getProvider() {
-        return "University of Rostock\nInstitute of Mathematics\nCITlab\nGundram Leifert\ngundram.leifert@uni-rostock.de";
+        return MetadataUtil.getProvider("Gundram Leifert", "gundram.leifert@planet-ai.de");
     }
 
     @Override