+++ src/main/java/de/uros/citlab/module/la/B2PSeam.java
@@ -33,7 +33,7 @@ import static org.opencv.core.Core.BORDER_DEFAULT;
  *
  * Since 29.01.2016
  *
- * @author Tobi <tobias.gruening@uni-rostock.de>
+ * @author Tobi <tobias.gruening@planet-ai.de>
  */
 public class B2PSeam extends ParamSetOrganizer implements IB2P {