+++ src/main/java/de/uros/citlab/module/norm/NormWritingInstance.java
@@ -53,7 +53,7 @@ public class NormWritingInstance implements IImageManipulator {
 
     @Override
     public String getProvider() {
-        return "University of Rostock\nInstitute of Mathematics\nCITlab\nGundram Leifert\ngundram.leifert@uni-rostock.de";
+        return MetadataUtil.getProvider("Gundram Leifert", "gundram.leifert@planet-ai.de");
     }
 
     @Override