+++ src/test/java/de/uros/citlab/module/TestWorkflow.java
@@ -23,8 +23,11 @@ import de.uros.citlab.module.baseline2polygon.Baseline2PolygonParser;
 import de.uros.citlab.module.htr.HTRParser;
 import de.uros.citlab.module.la.LayoutAnalysisParser;
 import de.uros.citlab.module.la.LayoutAnalysisURO_ML;
+import de.uros.citlab.module.types.Key;
+import de.uros.citlab.module.util.FileUtil;
 import de.uros.citlab.module.util.PageXmlUtil;
 import de.uros.citlab.module.util.PropertyUtil;
+import de.uros.citlab.module.util.TrainDataUtil;
 import de.uros.citlab.tokenizer.categorizer.CategorizerCharacterDft;
 import eu.transkribus.core.model.beans.pagecontent.PcGtsType;
 import eu.transkribus.core.util.PageXmlUtils;
@@ -35,26 +38,38 @@ import eu.transkribus.interfaces.types.Image;
 import org.apache.commons.io.FileUtils;
 import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
 import javax.xml.bind.JAXBException;
 import java.io.File;
 import java.io.IOException;
-import java.util.List;
-import java.util.Map;
-import java.util.Random;
+import java.util.*;
 
 public class TestWorkflow {
 
     public static Logger LOG = Logger.getLogger(TestWorkflow.class.getName());
 
-    Random r = new Random(1234);
+    static Random r = new Random(1234);
     private static File testFolder = new File(TestFiles.getPrefix(), "test_workflow_tmp/");
+    //    private static List<Pair<Image, String>> images=new LinkedList<>();
+    private static File fileLR = new File(testFolder, TrainDataUtil.lr);
 
-    @BeforeClass
-    public static void setUpClass() {
-        testFolder.mkdirs();
+    public static void createLR(String text) {
+//        try {
+//            for (File testFile : TestFiles.getTestFiles()) {
+//                images.add(new Pair<>(new Image(testFile.toURL()), testFile.getPath()));
+//            }
+//        } catch (IOException ex) {
+//            LOG.log(Logger.ERROR, ex);
+//            Assert.fail("creating test szenario did not work: (load image) " + ex.getMessage());
+//        }
+        List<String> words = Arrays.asList(text.split(" "));
+        List<String> out = new ArrayList<>(100);
+        for (int i = 0; i < 100; i++) {
+            out.add(String.join(" ", words));
+            Collections.shuffle(words,r);
+        }
+        FileUtil.writeLines(fileLR, out);
     }
 
     @AfterClass
@@ -73,6 +88,8 @@ public class TestWorkflow {
         IHtr htrParser = new HTRParser();
         IBaseline2Polygon b2pParser = new Baseline2PolygonParser(B2PSeamMultiOriented.class.getName());
         for (File testImgFile : TestFiles.getTestFiles()) {
+            String recoRef = PageXmlUtils.getFulltextFromLines(PageXmlUtil.unmarshal(PageXmlUtil.getXmlPath(testImgFile)));
+            createLR(recoRef);
             //SETUP
             String testFileImg = testImgFile.getName();
             String testFileImgBase = testFileImg.substring(0, testFileImg.lastIndexOf("."));
@@ -87,6 +104,7 @@ public class TestWorkflow {
             File testXmlFile_LA = null;
             File testXmlFile_LA_B2P = null;
             File testXmlFile_LA_HTR = null;
+            File testXmlFile_LA_HTR_LM = null;
             File testXmlFile_LA_B2P_HTR = null;
             PcGtsType createEmptyPcGtsType = null;
 //            try {
@@ -148,35 +166,44 @@ public class TestWorkflow {
             File htr = TestFiles.getHtrDft();
             try {
                 testXmlFile_LA_HTR = new File(testFolder, testFileImgBase + "_la_htr.xml");
+                testXmlFile_LA_HTR_LM = new File(testFolder, testFileImgBase + "_la_htr_lm.xml");
                 testXmlFile_LA_B2P_HTR = new File(testFolder, testFileImgBase + "_la_b2p_htr.xml");
                 FileUtils.copyFile(testXmlFile_LA, testXmlFile_LA_HTR);
+                FileUtils.copyFile(testXmlFile_LA, testXmlFile_LA_HTR_LM);
                 FileUtils.copyFile(testXmlFile_LA_B2P, testXmlFile_LA_B2P_HTR);
-                try {
-                    htrParser.process(htr.getAbsolutePath(),
-                            null,
-                            null,
-                            image,
-                            testXmlFile_LA_HTR.getAbsolutePath(),
-                            new File(testFolder, "confmats").getAbsolutePath(),
-                            null,
-                            PropertyUtil.setProperty(null, "raw", "true"));
-                    htrParser.process(htr.getAbsolutePath(),
-                            null,
-                            null,
-                            image,
-                            testXmlFile_LA_B2P_HTR.getAbsolutePath(),
-                            new File(testFolder, "confmats_b2p").getAbsolutePath(),
-                            null,
-                            PropertyUtil.setProperty(null, "raw", "true"));
-                } catch (Throwable e) {
-                    Assert.fail("applying HTR to image results an error: " + e.getMessage());
-                }
+//                try {
+                htrParser.process(htr.getAbsolutePath(),
+                        null,
+                        null,
+                        image,
+                        testXmlFile_LA_HTR_LM.getAbsolutePath(),
+                        new File(testFolder, "confmats_lm").getAbsolutePath(),
+                        null,
+                        PropertyUtil.setProperty(null, Key.HTR_USE_LR, fileLR.getAbsolutePath()));
+                htrParser.process(htr.getAbsolutePath(),
+                        null,
+                        null,
+                        image,
+                        testXmlFile_LA_HTR.getAbsolutePath(),
+                        new File(testFolder, "confmats").getAbsolutePath(),
+                        null,
+                        PropertyUtil.setProperty(null, "raw", "true"));
+                htrParser.process(htr.getAbsolutePath(),
+                        null,
+                        null,
+                        image,
+                        testXmlFile_LA_B2P_HTR.getAbsolutePath(),
+                        new File(testFolder, "confmats_b2p").getAbsolutePath(),
+                        null,
+                        PropertyUtil.setProperty(null, "raw", "true"));
+
+//                } catch (Throwable e) {
+//                    Assert.fail("applying HTR to image results an error: " + e.getMessage());
+//                }
             } catch (IOException ex) {
                 Assert.fail("creating test szenario did not work: (copy file for LA) " + ex.getMessage());
             }
             //checking quality between direct Polygons and Baseline->Poly-Alg ==> quality should not be too bad
-            String recoRef = PageXmlUtils.getFulltextFromLines(PageXmlUtil.unmarshal(PageXmlUtil.getXmlPath(testImgFile)));
-            double corRate, corRateB2P;
             {
                 String recoLaHtr = PageXmlUtils.getFulltextFromLines(PageXmlUtil.unmarshal(testXmlFile_LA_HTR));
                 ErrorModuleDynProg errorModule = new ErrorModuleDynProg(new CategorizerCharacterDft(), null, Boolean.FALSE);
@@ -184,10 +211,10 @@ public class TestWorkflow {
                 Map<Count, Long> map = errorModule.getCounter().getMap();
                 long gt = map.get(Count.GT);
                 long cor = map.get(Count.COR);
-                corRate = ((double) cor) / gt * 100;
+                double corRate = ((double) cor) / gt * 100;
 //                    System.out.println(map);
                 System.out.println(String.format("correct without B2P = %.2f%s", corRate, "%"));
-                Assert.assertEquals("correct rate is too bad without b2p", 100.0, corRate, 56.0);
+                Assert.assertEquals("correct rate changed without b2p and without lm", 47.28033472803347, corRate, 0.1);
             }
             {
                 String recoLaB2PHtr = PageXmlUtils.getFulltextFromLines(PageXmlUtil.unmarshal(testXmlFile_LA_B2P_HTR));
@@ -196,10 +223,22 @@ public class TestWorkflow {
                 Map<Count, Long> map = errorModule.getCounter().getMap();
                 long gt = map.get(Count.GT);
                 long cor = map.get(Count.COR);
-                corRateB2P = ((double) cor) / gt * 100;
+                double corRateB2P = ((double) cor) / gt * 100;
 //                    System.out.println(map);
                 System.out.println(String.format("correct with B2P = %.2f%s", corRateB2P, "%"));
-                Assert.assertEquals("correct rate is too bad with b2p", 100.0, corRate, 60.0);
+                Assert.assertEquals("correct rate changed with b2p and without lm", 46.23430962343097, corRateB2P, 0.1);
+            }
+            {
+                String reco = PageXmlUtils.getFulltextFromLines(PageXmlUtil.unmarshal(testXmlFile_LA_HTR_LM));
+                ErrorModuleDynProg errorModule = new ErrorModuleDynProg(new CategorizerCharacterDft(), null, Boolean.FALSE);
+                errorModule.calculate(reco, recoRef);
+                Map<Count, Long> map = errorModule.getCounter().getMap();
+                long gt = map.get(Count.GT);
+                long cor = map.get(Count.COR);
+                double corRateB2P = ((double) cor) / gt * 100;
+//                    System.out.println(map);
+                System.out.println(String.format("correct with B2P and with n-gram = %.2f%s", corRateB2P, "%"));
+                Assert.assertEquals("correct rate changed without b2p and with n-gram language model", 55.85774058577406, corRateB2P, 0.1);
             }
 //            Assert.assertEquals("htr should benefit from B2P", corRate + Math.abs(corRate - corRateB2P), corRateB2P, Math.abs(corRate - corRateB2P));
         }